Change classes on modules for forms
