+++ app/forms/articles/update_form.rb
@@ -1,14 +1,7 @@
 # frozen_string_literal: true
 
-# module Article
-  class Article::UpdateForm
-    include ActiveModel::Model
-
-    attr_accessor :content, :category, :article_tags, :user, :article
-
-    validates :category, :content, :article_tags, presence: true
-    validate :user_is_admin?, :tag_ids_exists?
-
+module Articles
+  class UpdateForm < Articles::BaseForm
     def initialize(params, user, article)
       @article = article
       @user = user
@@ -29,14 +22,6 @@
 
     private
 
-    def user_is_admin?
-      errors.add(:base, 'You are not authorised for this action') unless user.is_a?(Admin)
-    end
-
-    def tag_ids_exists?
-      errors.add(:base, 'Tag ids invalid') unless article_tags[:tag_ids].all? { |tag_id| Tag.find_by(id: tag_id) }
-    end
-
     def update_article_tags!
       article.article_tags.destroy_all
 
@@ -47,4 +32,4 @@
       article.update!(category: category, content: content)
     end
   end
-# end
+end
