+++ spec/factories/article_factory.rb
@@ -1,5 +1,18 @@
 # frozen_string_literal: true
 
+# == Schema Information
+#
+# Table name: articles
+#
+#  id           :bigint           not null, primary key
+#  category     :string           not null
+#  tsv_category :tsvector
+#  tsv_content  :tsvector
+#  status       :string           default("draft"), not null
+#  user_id      :bigint           not null
+#  created_at   :datetime         not null
+#  updated_at   :datetime         not null
+#
 FactoryBot.define do
   factory :article do
     transient do
