+++ spec/forms/articles/update_form_spec.rb
@@ -2,10 +2,12 @@
 
 require 'rails_helper'
 
-RSpec.describe Article::CreateForm do
-  subject(:form) { described_class.new(params, current_admin).save }
+RSpec.describe Articles::UpdateForm do
+  subject(:form) { described_class.new(params, current_admin, article).save }
 
   let(:current_admin) { create(:admin) }
+  let!(:article) { create(:article, user_id: current_admin.id) }
+
   let(:tags) { create_list(:tag, 2) }
   let(:content) { FFaker::Lorem.paragraph }
   let(:category) { FFaker::Lorem.word }
@@ -16,18 +18,18 @@ RSpec.describe Article::CreateForm do
         {
           content: content,
           category: category,
-          article_tags: { tag_ids: tags.pluck(:id) }
+          article_tags: { tag_ids: article.tags.pluck(:id) + tags.pluck(:id) }
         }
       end
 
-      it 'changes Articles count' do
-        expect { form }.to change(Article, :count).from(0).to(1)
+      it 'not changes Articles count' do
+        expect { form }.not_to change(Article, :count)
       end
 
       it 'changes ArticleTags count' do
-        expect { form }.to change(ArticleTag, :count).from(0).to(2)
+        expect { form }.to change(ArticleTag, :count).from(2).to(4)
 
-        expect(current_admin.articles.first.article_tags.count).to eq(2)
+        expect(current_admin.articles.first.article_tags.count).to eq(4)
       end
 
       it 'creates a new instance of Article with correct values' do
@@ -49,7 +51,9 @@ RSpec.describe Article::CreateForm do
 
   describe 'fail' do
     context 'when user is reader' do
-      let!(:current_admin) { create(:reader) }
+      let(:current_admin) { create(:reader) }
+      let(:admin) { create(:admin) }
+      let!(:article) { create(:article, user_id: admin.id) }
 
       let(:params) do
         {
@@ -117,7 +121,7 @@ RSpec.describe Article::CreateForm do
     end
 
     context 'when raises error' do
-      before { allow(Article).to receive(:create!).and_raise(ActiveRecord::RecordInvalid) }
+      before { allow(article).to receive(:update!).and_raise(ActiveRecord::RecordInvalid) }
 
       let(:params) do
         {
